+++ test_runner/src/main/kotlin/ftl/gc/GcAndroidTestMatrix.kt
@@ -3,23 +3,21 @@ package ftl.gc
 import com.google.api.services.testing.Testing
 import com.google.api.services.testing.model.Account
 import com.google.api.services.testing.model.AndroidDeviceList
-import com.google.api.services.testing.model.AndroidInstrumentationTest
-import com.google.api.services.testing.model.Apk
 import com.google.api.services.testing.model.ClientInfo
-import com.google.api.services.testing.model.DeviceFile
 import com.google.api.services.testing.model.EnvironmentMatrix
 import com.google.api.services.testing.model.EnvironmentVariable
-import com.google.api.services.testing.model.FileReference
 import com.google.api.services.testing.model.GoogleAuto
 import com.google.api.services.testing.model.GoogleCloudStorage
-import com.google.api.services.testing.model.RegularFile
 import com.google.api.services.testing.model.ResultStorage
 import com.google.api.services.testing.model.TestMatrix
 import com.google.api.services.testing.model.TestSetup
 import com.google.api.services.testing.model.TestSpecification
 import com.google.api.services.testing.model.ToolResultsHistory
 import ftl.args.AndroidArgs
-import ftl.args.ShardChunks
+import ftl.gc.android.mapGcsPathsToApks
+import ftl.gc.android.mapToDeviceFiles
+import ftl.gc.android.setupAndroidTest
+import ftl.run.platform.android.AndroidTestConfig
 import ftl.util.join
 import ftl.util.timeoutToSeconds
 
@@ -30,13 +28,12 @@ object GcAndroidTestMatrix {
         value = this@toEnvironmentVariable.value
     }
 
+    @Suppress("LongParameterList")
     fun build(
-        appApkGcsPath: String,
-        testApkGcsPath: String,
+        androidTestConfig: AndroidTestConfig,
         otherFiles: Map<String, String>,
         runGcsPath: String,
         androidDeviceList: AndroidDeviceList,
-        testShards: ShardChunks,
         args: AndroidArgs,
         toolResultsHistory: ToolResultsHistory,
         additionalApkGcsPaths: List<String>
@@ -48,19 +45,6 @@ object GcAndroidTestMatrix {
             .setClientInfoDetails(args.clientDetails?.toClientInfoDetailList())
 
         val matrixGcsPath = join(args.resultsBucket, runGcsPath)
-
-        val androidInstrumentation = AndroidInstrumentationTest()
-            .setAppApk(FileReference().setGcsPath(appApkGcsPath))
-            .setTestApk(FileReference().setGcsPath(testApkGcsPath))
-            .setupTestTargets(args, testShards)
-
-        if (args.testRunnerClass != null) {
-            androidInstrumentation.testRunnerClass = args.testRunnerClass
-        }
-
-        if (args.useOrchestrator) {
-            androidInstrumentation.orchestratorOption = "USE_ORCHESTRATOR"
-        }
 
         // --auto-google-login
         // https://cloud.google.com/sdk/gcloud/reference/firebase/test/android/run
@@ -86,11 +70,11 @@ object GcAndroidTestMatrix {
         val testTimeoutSeconds = timeoutToSeconds(args.testTimeout)
 
         val testSpecification = TestSpecification()
-            .setAndroidInstrumentationTest(androidInstrumentation)
             .setDisablePerformanceMetrics(!args.performanceMetrics)
             .setDisableVideoRecording(!args.recordVideo)
             .setTestTimeout("${testTimeoutSeconds}s")
             .setTestSetup(testSetup)
+            .setupAndroidTest(androidTestConfig)
 
         val resultsStorage = ResultStorage()
             .setGoogleCloudStorage(GoogleCloudStorage().setGcsPath(matrixGcsPath))
@@ -111,16 +95,4 @@ object GcAndroidTestMatrix {
             throw RuntimeException(e)
         }
     }
-}
-
-private fun List<String>?.mapGcsPathsToApks(): List<Apk>? = this
-    ?.takeIf { it.isNotEmpty() }
-    ?.map { gcsPath -> Apk().setLocation(FileReference().setGcsPath(gcsPath)) }
-
-private fun Map<String, String>.mapToDeviceFiles() = map { (devicePath: String, gcsFilePath: String) ->
-    DeviceFile().setRegularFile(
-        RegularFile()
-            .setDevicePath(devicePath)
-            .setContent(FileReference().setGcsPath(gcsFilePath))
-    )
 }