+++ test_runner/src/main/kotlin/ftl/run/platform/android/CreateInstrumentationConfig.kt
@@ -1,0 +1,18 @@
+package ftl.run.platform.android
+
+import ftl.args.AndroidArgs
+import ftl.args.ShardChunks
+import ftl.args.yml.UploadedApks
+
+internal fun AndroidArgs.createInstrumentationConfig(
+    uploadedApks: UploadedApks,
+    testShards: ShardChunks
+) = AndroidTestConfig.Instrumentation(
+    appApkGcsPath = uploadedApks.app,
+    testApkGcsPath = uploadedApks.test!!,
+    testRunnerClass = testRunnerClass,
+    orchestratorOption = "USE_ORCHESTRATOR".takeIf { useOrchestrator },
+    disableSharding = disableSharding,
+    numUniformShards = numUniformShards,
+    testShards = testShards
+)