+++ test_runner/src/test/kotlin/ftl/gc/android/CreateAndroidInstrumentationTestTest.kt
@@ -1,11 +1,8 @@
-package ftl.gc
+package ftl.gc.android
 
 import com.google.api.services.testing.model.AndroidInstrumentationTest
 import com.google.api.services.testing.model.FileReference
-import ftl.args.AndroidArgs
 import ftl.args.ShardChunks
-import io.mockk.every
-import io.mockk.mockk
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull
 import org.junit.Test
@@ -15,13 +12,15 @@ class UtilsKtTest {
     @Test
     fun `setupTestTargets should setup testTargets`() {
         // given
-        val args = mockk<AndroidArgs> {
-            every { disableSharding } returns true
-        }
         val testShards: ShardChunks = emptyList()
 
         // when
-        val actual = AndroidInstrumentationTest().setupTestTargets(args, testShards)
+        val actual = AndroidInstrumentationTest()
+            .setupTestTargets(
+                disableSharding = true,
+                testShards = testShards,
+                numUniformShards = null
+            )
             .testTargets
 
         // then
@@ -32,16 +31,16 @@ class UtilsKtTest {
     fun `setupTestTargets should setup uniformSharding`() {
         // given
         val expectedTestTargets = emptyList<String>()
-        val args = mockk<AndroidArgs> {
-            every { disableSharding } returns false
-            every { numUniformShards } returns 50
-        }
         val testShards: ShardChunks = listOf(expectedTestTargets)
 
         // when
         val actual = AndroidInstrumentationTest()
             .setTestApk(FileReference().setGcsPath("testApk"))
-            .setupTestTargets(args, testShards)
+            .setupTestTargets(
+                disableSharding = false,
+                testShards = testShards,
+                numUniformShards = 50
+            )
 
         // then
         assertEquals(0, actual.shardingOption.uniformSharding.numShards)
@@ -52,13 +51,14 @@ class UtilsKtTest {
     fun `setupTestTargets should setup manualSharding`() {
         // given
         val shardChunks: ShardChunks = listOf(emptyList(), emptyList())
-        val args = mockk<AndroidArgs> {
-            every { disableSharding } returns false
-            every { numUniformShards } returns null
-        }
 
         // when
-        val actual = AndroidInstrumentationTest().setupTestTargets(args, shardChunks)
+        val actual = AndroidInstrumentationTest()
+            .setupTestTargets(
+                disableSharding = false,
+                testShards = shardChunks,
+                numUniformShards = null
+            )
             .shardingOption
             .manualSharding
             .testTargetsForShard